Add coverage thresholds
Overall, I would like the coverage to stay at 100% all the time. In
practice, this may not be achievable, but 80% should be reasonable. The
exception to this is the core business logic, which should have 100%
coverage all the time given its importance. Hence, 100% coverage is
required for anything in the model directory.