Add support for indexing A and B questions
Starting with question 16, if a question is answered incorrectly, the
next questions asked are not the next numerical question (e.g., if
question 5 is erred, question 6 is asked next), but we move to A and B
questions instead (e.g. if question 16 is erred, question 16A is next).

Add a class to convert between string versions of the raw question
numbers and the numerical index that corresponds to that question. In a
quiz, some indexes might be skipped, but they have definitive indexes so
that a given index is _always_ a given question number.