Add correct answer tracking for teams
Add score tracking for correct answers to teams. Points for a quiz out
with a bonus are not applied to the team as it's only an individual
bonus.

Deductions for consecutive errors by a quizzer are not yet added. Single
errors by a quizzer don't deduct points, and a team is allowed up to 2
errors without points.