Change quizzers list in Team to array
To simplify the memory structure of the Team and make it easier to
return our list of quizzers, change the Map that was holding the
quizzers to an array and update method implementations accordingly. This
does not change the API of the Team nor does it change any external
functionality.